Windows/MSVC build: source portability fixes (_WIN32/_MSC_VER guards)
Apply Windows/MSVC portability fixes across the core sources, guarded by
_WIN32/_MSC_VER so POSIX/GCC/Clang paths are unchanged:
- Guard POSIX-only includes (unistd.h, sys/mman.h, pthread, ...) and
  provide Win32 equivalents (TempFilePath/TempDirectoryPath via Win32
  file APIs; HierarchicalTimer cpuNow via GetThreadTimes).
- MSVC C4716 no-return-stub pragmas / explicit returns.
- Include the folly::available_concurrency shim at core call sites where
  the pinned folly only provides hardware_concurrency().
- int128 shim support: free operator+/- for built-in-left operands,
  explicit narrowing casts (Sequence, DecimalFunctions ceil), and
  is_arithmetic||int128 enable_if plus folly::dynamic handling in the
  fuzzer constrained generators.
- Use const char* error_source/error_code constants (drop .c_str()).
- Carry over the proven base fixes: std::construct_at size_t->int32_t
  narrowing, eager folly::hasher<ComplexType> instantiation guards, and
  move-only struct copy in maps.

+++ velox/type/TimestampConversion.cpp
@@ -864,10 +864,10 @@ int32_t extractISODayOfTheWeek(int64_t daysSinceEpoch) {
   // 7  = 4
   if (daysSinceEpoch < 0) {
     // negative date: start off at 4 and cycle downwards
-    return (7 - ((-int128_t(daysSinceEpoch) + 3) % 7));
+    return static_cast<int32_t>(7 - ((-int128_t(daysSinceEpoch) + 3) % 7));
   } else {
     // positive date: start off at 4 and cycle upwards
-    return ((int128_t(daysSinceEpoch) + 3) % 7) + 1;
+    return static_cast<int32_t>(((int128_t(daysSinceEpoch) + 3) % 7) + 1);
   }
 }
 
@@ -992,7 +992,7 @@ Timestamp fromParsedTimestampWithTimeZone(
   } else if (parsed.offsetMillis.has_value()) {
     auto seconds = parsed.timestamp.getSeconds();
     // use int128_t to avoid overflow.
-    __int128_t nanos = parsed.timestamp.getNanos();
+    int128_t nanos = parsed.timestamp.getNanos();
     seconds -= parsed.offsetMillis.value() / util::kMillisPerSecond;
     nanos -= (parsed.offsetMillis.value() % util::kMillisPerSecond) *
         util::kNanosPerMicro * util::kMicrosPerMsec;
@@ -1003,7 +1003,7 @@ Timestamp fromParsedTimestampWithTimeZone(
       seconds += 1;
       nanos -= Timestamp::kNanosInSecond;
     }
-    parsed.timestamp = Timestamp(seconds, nanos);
+    parsed.timestamp = Timestamp(seconds, static_cast<uint64_t>(nanos));
   } else {
     if (sessionTimeZone) {
       parsed.timestamp.toGMT(*sessionTimeZone);
